% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie-internals.R
\name{Tweedie internals}
\alias{Tweedie internals}
\alias{dtweedie_dldphi_saddle}
\alias{dtweedie_logv_bigp}
\alias{dtweedie_logw_smallp}
\alias{dtweedie_interp}
\alias{dtweedie_jw_smallp}
\alias{dtweedie_kv_bigp}
\alias{dtweedie_Fortran}
\alias{dtweedie_Inversion_Report}
\alias{dtweedie_Inversion_Threemethods}
\alias{dtweedie_dlogfdphi}
\alias{dtweedie_logl}
\alias{dtweedie_dldphi}
\alias{dtweedie_logl_saddle}
\alias{dtweedie_series_bigp}
\alias{dtweedie_series_smallp}
\alias{logLiktweedie}
\alias{stored_grids}
\alias{twcomputation}
\alias{sort_notation}
\alias{check_inputs}
\alias{special_cases}
\title{Tweedie internal function}
\usage{
dtweedie_dlogfdphi(y, mu, phi, power)
dtweedie_logl(phi, y, mu, power)
dtweedie_logl_saddle( phi, power, y, mu, eps=0)
dtweedie_logv_bigp( y, phi, power)
dtweedie_logw_smallp(y, phi, power)
dtweedie_interp(grid, nx, np, xix.lo, xix.hi,p.lo, p.hi, power, xix)
dtweedie_jw_smallp(y, phi, power )
dtweedie_kv_bigp(y, phi, power)
dtweedie_series_bigp(power, y, mu, phi)
dtweedie_series_smallp(power, y, mu, phi)
stored_grids(power)
check_inputs(y, mu, phi, power, type = "standard")
sort_notation(xi = NULL, power = NULL)
special_cases(y, mu, phi, power, type="PDF", verbose = FALSE)
}
\arguments{
\item{phi}{the dispersion \eqn{\phi}{phi}.}

\item{mu}{the mean parameter \eqn{\mu}{mu}.}

\item{power}{the value of \eqn{p}{power} such that the variance is \eqn{\mbox{var}[Y]=\phi\mu^p}{var(Y) = phi * mu^power}.}

\item{y}{the vector of responses.}

\item{xi}{a synonym for \code{power}.}

\item{grid}{the interpolation grid necessary for the given value of \eqn{p}{power}.}

\item{nx}{the number of interpolation points in the \eqn{\xi}{xi} dimension.}

\item{np}{the number of interpolation points in the \eqn{p}{power}-dimension.}

\item{xix.lo}{the lower value of the transformed \eqn{\xi}{xi} value used in the interpolation grid. (Note that the value of \eqn{\xi}{xi} is from \eqn{0} to \eqn{\infty}{infty}, and is transformed such that it is on the range \eqn{0} to \eqn{1}.)}

\item{xix.hi}{the higher value of the transformed \eqn{\xi}{xi} value used in the interpolation grid.}

\item{p.lo}{the lower value of the \eqn{p} value used in the interpolation grid.}

\item{p.hi}{the higher value of the \eqn{p} value used in the interpolation grid.}

\item{xix}{the value of the transformed \eqn{\xi}{xi} at which a value is sought.}

\item{verbose}{logical; if \code{TRUE}, some details of the algorithm are returned.}

\item{eps}{the offset in computing the variance function in the saddlepoint approximation. The default is \code{eps=1/6} (as suggested by Nelder and Pregibon, 1987).}

\item{type}{in \code{check_inputs}, the type of function for which inputs should be checked (one of \code{"standard"} (for \code{dtweedie} and \code{ptweedie}; the default), \code{"random"} (for \code{rtweedie}) or \code{"quantile"} (for \code{qtweedie})); in \code{special_cases}, one of \code{"PDF"} (the probability density function; the default) or \code{"CDF"} (the cumulative distribution function).}
}
\description{
Internal tweedie functions. \bold{These are not to be called by the user.}
}
\details{
Tweedie internal functions
}
\references{
Dunn, P. K. and Smyth, G. K. (2008).
Evaluation of Tweedie exponential dispersion model densities by Fourier inversion.
\emph{Statistics and Computing}, 
\bold{18}, 73--86.
\doi{10.1007/s11222-007-9039-6}

Dunn, Peter K and Smyth, Gordon K (2005).
Series evaluation of Tweedie exponential dispersion model densities
\emph{Statistics and Computing},
\bold{15}(4). 267--280.
\doi{10.1007/s11222-005-4070-y}

Jorgensen, B. (1997).
\emph{Theory of Dispersion Models}.
Chapman and Hall, London.

	Nelder, J. A. and Pregibon, D. (1987).
	An extended quasi-likelihood function
	\emph{Biometrika},
	\bold{74}(2), 221--232.
	\doi{10.1093/biomet/74.2.221}
}
\author{
Peter Dunn (\email{pdunn2@usc.edu.au})
}
\keyword{internal}
