% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtweedie_inversion.R
\name{dtweedie_inversion}
\alias{dtweedie_inversion}
\alias{dtweedie.inversion}
\title{Fourier Inversion Evaluation for the Tweedie Probability Function}
\usage{
dtweedie_inversion(y, power, mu, phi, method = 3, verbose = FALSE, details = FALSE)
}
\arguments{
\item{y}{vector of quantiles.}

\item{power}{scalar; the power parameter \eqn{p}{power}.}

\item{mu}{the mean parameter \eqn{\mu}{mu}.}

\item{phi}{the dispersion parameter \eqn{\phi}{phi}.}

\item{method}{the method to use; one of \code{1}, \code{2}, or \code{3} (the default).}

\item{verbose}{logical; if \code{TRUE}, display some internal computation details. The default is \code{FALSE}.}

\item{details}{logical; if \code{TRUE}, return a list with basic details of the integration. The default is \code{FALSE}.}
}
\value{
A numeric vector of densities if \code{details=FALSE}; if \code{details=TRUE}, return a list with \code{density} (the density values), \code{regions} (the number of integration regions used) and \code{methods} (which of the three methods was used).

For special cases of \eqn{p} (i.e., \eqn{p = 0, 1, 2, 3}), where no inversion is needed, \code{regions} and \code{method} are set to \code{NA} for all values of \code{y}.
For special cases of \code{y} for other values of \eqn{p} (i.e., \eqn{P(Y = 0)}), \code{regions} and \code{method} are set to \code{NA}.
}
\description{
Evaluates the probability density function (\acronym{pdf}) for Tweedie distributions using Fourier inversion, 
for given values of the dependent variable \code{y}, the mean \code{mu}, dispersion \code{phi}, and power parameter \code{power}.
\emph{Not usually called by general users}, but can be used in the case of evaluation problems.
}
\note{
The three methods are described in Dunn & Smyth (2008).
}
\examples{
# Plot a Tweedie density
y <- seq(0, 5, length = 100)
fy <- dtweedie_inversion(y, power = 1.1, mu = 1, phi = 1)
plot(y, fy, type = "l", lwd = 2, ylab = "Density")

}
\references{
Dunn, P. K. and Smyth, G. K. (2008).
Evaluation of Tweedie exponential dispersion model densities by Fourier inversion.
\emph{Statistics and Computing}, 
\bold{18}, 73--86.
\doi{10.1007/s11222-007-9039-6}

Dunn, P. K. and Smyth, G. K. (2008).
Evaluation of Tweedie exponential dispersion model densities by Fourier inversion.
\emph{Statistics and Computing}, 
\bold{18}, 73--86.
\doi{10.1007/s11222-007-9039-6}
}
\keyword{distribution}
